% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_function.R
\name{company_function}
\alias{company_function}
\title{Obtain consumer complaint information about a specific company}
\usage{
company_function(product_option = TRUE, issue_option = TRUE,
  response_option = TRUE, public_option = FALSE)
}
\arguments{
\item{product_option}{A logical indicating whether to aggregate all the consumer complaints by product category under the selected company. If (\code{TRUE}), such information will be created in the form of a dataframe as the output. Defaults to (\code{TRUE})}

\item{issue_option}{A logical indicating whether to aggregate all the consumer complaints by issue category under the selected company. If (\code{TRUE}), such information will be created in the form of a dataframe as the output. Defaults to (\code{TRUE})}

\item{response_option}{A logical indicating whether to aggregate all the consumer complaints by company response category under the selected company. If (\code{TRUE}), such information will be created in the form of a dataframe as the output. Defaults to (\code{TRUE})}

\item{public_option}{A logical indicating whether to aggregate all the consumer complaints by public company response category under the selected company. If (\code{TRUE}), such information will be created in the form of a dataframe as the output. Defaults to (\code{FALSE}), as not all companies have a public response}
}
\value{
A list of the dataframe(s) of the options selected. Multiple dataframes are presented as a list because a function can only return one output. As such, by feeding multiple dataframes through a list, multiple outputs can be achieved.
}
\description{
It displays various useful summaries of the number of consumer complaints pertaining to different categories of interest for a single company, and puts them in dataframe(s).
}
\details{
The function will first prompt the user to enter the desired company as an input for its first parameter. The input will be used to then guess possible suggestions related to that company name. This is necessary as the company name has to be inputted exactly the same as what is recorded in the consumer complaint database. Once the user finds the right suggestion, the user is prompted again to input the correct name of the company. Once that is complete, it will display any of the four possible summarization options. The user can specify whether to include any one or all of the summarization options by specifying (\code{TRUE}) or (\code{FALSE}).
}
\examples{
#No examples can be demonstrated, as it requires a manual user input from a prompt
}
\author{
Jae Woong Ham
}
