% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{general_function}
\alias{general_function}
\title{Obtain consumer complaint information in a flexible search}
\usage{
general_function(company = "BANK OF AMERICA, NATIONAL ASSOCIATION",
  product = "Checking or savings account",
  issue = "Managing an account", limit = 10)
}
\arguments{
\item{company}{A character input from the user to specify a company name. Defaults to Bank of America.}

\item{product}{A character input from the user to specify a product category. Defaults to checking or savings account category.}

\item{issue}{A character input from the user to specify an issue category. Defaults to managing an account.}

\item{limit}{A integer input from the user to limit the number of results. Defaults to 10.}
}
\value{
a dataframe of the chosen parameters
}
\description{
It displays all associated datapoint(s) depending on what was specified by the user. The user has varying options to specify what kind of complaint the user wants. The output is a single dataframe that contains all the relevant datapoints.
}
\details{
The function will take the input parameters from the user, and the input paramters need to be correctly matched with how it is recorded in the database, unlike other functions in this package.
}
\examples{
general_function()
}
\author{
Jae Woong Ham
}
