% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_state_function.R
\name{heatmap_state_function}
\alias{heatmap_state_function}
\title{Visualization tool for number of complaints by state.}
\usage{
heatmap_state_function(df = sampledf)
}
\arguments{
\item{df}{Dataframe (two columns)}
}
\value{
Graph of the United States with the ascribed labels
}
\description{
Using a heatmap, it displays the various frequencies of complaints by state on the US map.
}
\details{
The function can only take the specific dataframes generated by the other functions in this package that contains the state information. Moreover, the data can only contain two columns, whereas the first column is the number of states and the second column is the number of complaints.
}
\examples{
heatmap_state_function()
}
\author{
Jae Woong Ham
}
