% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tool_function.R
\name{visualization_tool_function}
\alias{visualization_tool_function}
\title{Visualization tool for number of complaints by any category.}
\usage{
visualization_tool_function(df = sampledf, limit = 10)
}
\arguments{
\item{df}{Dataframe (two columns)}

\item{limit}{Places a limit on the output of the barplot in terms of number of datapoints/ observations}
}
\value{
Barplot with the same labels and values from the original dataframe
}
\description{
It displays a rotated barplot that features the categorical variables on the y axis, and values for the number of count of complaints on the x axis.
}
\details{
The function can only take the specific dataframes generated by the other functions in this package that has two columns. The first column can feature any categorical labels, while the second column needs to contain values. The function will then take the dataframe and visualize the contents on a tailored barplot.
}
\examples{
visualization_tool_function()
}
\author{
Jae Woong Ham
}
